+++ app/src/main/java/reikai/presentation/novel/reader/NovelReaderScreenModel.kt
@@ -167,6 +167,7 @@ class NovelReaderScreenModel(
         novelRepo = novelRepo,
         sourceManager = sourceManager,
         installer = installer,
+        preferences = novelPreferences,
         readDownloaded = { novel, chapter -> downloadManager.getChapterText(novel, chapter) },
     )